优化文件结构
